Refactor grid from model
